ai_architect_backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.0
pydantic-settings>=2.10.0
pymongo>=4.12.0
python-dotenv>=1.1.0
openai>=1.86.0
anthropic>=0.55.0
